requirements.txt
# requirements.txt

requests
beautifulsoup4
dnspython
PyYAML
